requirements.txt
fastmcp
python-dotenv
ipykernel
httpx
yfinance
liteLLM
openai-agents[liteLLM]